misc/HtmlTestSuite/results/test-1288.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>Fragment directive API</title>
<meta charset="utf-8"/>
<link href="https://wicg.github.io/ScrollToTextFragment/" rel="help"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/resources/testdriver.js"/>
<script src="/resources/testdriver-vendor.js"/>
<script>
test(t =&gt; {
  assert_equals(typeof(document.fragmentDirective), 'object', 'document.fragmentDirective is defined');
}, 'Scroll to text is feature detectable via document.fragmentDirective');

test(t =&gt;{
  document.fragmentDirective = 'text=test';
  assert_equals(window.scrollY, 0, 'Setting document.fragmentDirective did not have an effect on scroll position');
  assert_equals(typeof(document.fragmentDirective), 'object', 'document.fragmentDirective is still an object type');
  assert_equals(Object.keys(document.fragmentDirective).length, 0, 'document.fragmentDirective has no properties');
}, 'Setting document.fragmentDirective has no effect');
</script>
<style>
  body {
    height: 3200px;
  }
  #text {
    position: absolute;
    top: 3000px;
  }
</style>
</head><body>
  <p id="text">This is a test page</p>
</body>
</html>
